rafctored fiels for clarity and documentation
