+++ Makefile
@@ -1,6 +1,8 @@
 .PHONY: cli-build build install clean
 
-default: cli-build
+default: all
+
+all: getdeps fmt cli-build install
 
 fmt:
 	./scripts/make.sh fmt	
@@ -8,7 +10,12 @@ fmt:
 cli-build:
 	./scripts/make.sh binary
 
+cli-getdeps:
+	./scripts/make.sh getdeps
+
 build: cli-build
+
+getdeps: cli-getdeps
 
 install: build
 	./scripts/make.sh install
