+++ api/handler/consul/README.md
@@ -1,0 +1,5 @@
+CONSUL
+======
+
+Connect to a Consul Agent to provide Operations based on
+service discovery and key value management.
