+++ api/handler/consul/api.go
@@ -1,0 +1,1 @@
+package consul
