--- api/handler/consul/consul.go
@@ -1,6 +1,0 @@
-package consul
-
-/**
- * Connect to a Consul Agent to provide Operations based on
- * service discovery and key value management.
- */
