+++ api/handler/kontena/README.md
@@ -1,0 +1,5 @@
+Kontena
+=======
+
+Provide Kontena based orchestration through
+a custom Handler.
