--- api/handler/kontena/kontena.go
@@ -1,6 +1,0 @@
-package kontena
-
-/**
- * Provide Kontena based orchestration through
- * a custom Handler.
- */
