+++ api/handler/local/README.md
@@ -1,0 +1,6 @@
+LOCAL
+=====
+
+Provide a Handler, and Operations, based on a local project, meaning
+based on current local path, and the configuration files contained
+therein (and perhaps also in the user home folder somewhere)
