--- api/handler/local/local.go
@@ -1,7 +1,0 @@
-package local
-
-/**
- * Provide a Handler, and Operations, based on a local project, meaning
- * based on current local path, and the configuration files contained
- * therein (and perhaps also in the user home folder somewhere)
- */
