+++ api/handler/null/README.md
@@ -1,0 +1,7 @@
+NULL
+----
+
+The null implementation centers around the null handler, which
+provides Operation implementations that do nothing, and are
+generally safe to use in cases where an Operation is kind of
+expected, but you don't have a handler to fit the role.
