+++ api/handler/rsa/README.md
@@ -1,0 +1,5 @@
+RSA
+===
+
+RSA adds a handler which provides primarily security authentication
+operations, via RSA application on passed configurations.
