--- api/handler/rsa/rsa.go
@@ -1,7 +1,0 @@
-package rsa
-
-/**
- * RSA adds a handler which provides primarily security authentication
- * operations, via RSA application on passed configurations.
- *
- */
