+++ api/handler/upcloud/README.md
@@ -1,0 +1,6 @@
+UPCLOUD 
+=======
+
+Provide a handler that connects to UpCloud and defines
+Operations for orchestration and monitoring for UpCloud
+servers and apps
