--- api/handler/upcloud/handler.go
@@ -1,7 +1,0 @@
-package upcloud
-
-/**
- * Provide a handler that connects to UpCloud and defines
- * Operations for orchestration and monitoring for UpCloud
- * servers and apps
- */
