+++ api/operation/app/app.go
@@ -1,0 +1,5 @@
+package api
+
+/**
+ * API.api operations deal directly with the API itself
+ */
