+++ api/operation/orchestrate/README.md
@@ -1,0 +1,13 @@
+ORCHESTRATE 
+===========
+
+Orchestration operations are meant to control the state
+of the app.
+
+Orchestration is analogous to either swarm, compose or
+direct Docker control operations, and typically centers
+around Up, Down, Start, Stop etc.
+Additional orchestration Operations should be used for
+release, branching and forking concepts.
+
+## QUESTION should branching/releasing be it's own concept?
