+++ scripts/binary
@@ -9,7 +9,7 @@ go generate
 BUILDTIME=$(date --rfc-3339 ns | sed -e 's/ /T/') &> /dev/null
 GITCOMMIT=$(git rev-parse --short HEAD)
 
-# Build binaries
+# Build cli binaries
 go build \
    -ldflags="-w -X ${WUNDERTOOLS_PKG}/version.GITCOMMIT=${GITCOMMIT} -X ${WUNDERTOOLS_PKG}/version.BUILDTIME=${BUILDTIME} -X ${WUNDERTOOLS_PKG}/version.SHOWWARNING=${SHOWWARNING}" \
    -o "${WUNDERTOOLS_BUILD_BINARY_PATH}" \
